+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/Brevoppretter.kt
@@ -3,7 +3,6 @@ package no.nav.etterlatte.brev
 import no.nav.etterlatte.brev.adresse.AdresseService
 import no.nav.etterlatte.brev.behandling.PersonerISak
 import no.nav.etterlatte.brev.db.BrevRepository
-import no.nav.etterlatte.brev.hentinformasjon.behandling.BehandlingService
 import no.nav.etterlatte.brev.model.Adresse
 import no.nav.etterlatte.brev.model.Brev
 import no.nav.etterlatte.brev.model.BrevDataRedigerbar
@@ -12,7 +11,6 @@ import no.nav.etterlatte.brev.model.BrevkodeRequest
 import no.nav.etterlatte.brev.model.Mottaker
 import no.nav.etterlatte.brev.model.OpprettNyttBrev
 import no.nav.etterlatte.libs.common.behandling.SakType
-import no.nav.etterlatte.libs.common.feilhaandtering.UgyldigForespoerselException
 import no.nav.etterlatte.libs.common.person.Folkeregisteridentifikator
 import no.nav.etterlatte.libs.common.person.MottakerFoedselsnummer
 import no.nav.etterlatte.libs.common.person.UkjentVergemaal
@@ -20,43 +18,13 @@ import no.nav.etterlatte.libs.common.person.Vergemaal
 import no.nav.etterlatte.libs.common.sak.SakId
 import no.nav.etterlatte.libs.common.tidspunkt.Tidspunkt
 import no.nav.etterlatte.libs.ktor.token.BrukerTokenInfo
-import no.nav.etterlatte.libs.ktor.token.Saksbehandler
 import java.util.UUID
 
 class Brevoppretter(
     private val adresseService: AdresseService,
     private val db: BrevRepository,
-    private val behandlingService: BehandlingService,
     private val innholdTilRedigerbartBrevHenter: InnholdTilRedigerbartBrevHenter,
 ) {
-    suspend fun opprettVedtaksbrev(
-        sakId: SakId,
-        behandlingId: UUID,
-        brukerTokenInfo: BrukerTokenInfo,
-        brevKodeMapping: (b: BrevkodeRequest) -> Brevkoder,
-        brevDataMapping: suspend (BrevDataRedigerbarRequest) -> BrevDataRedigerbar,
-    ): Brev {
-        require(db.hentBrevForBehandling(behandlingId, Brevtype.VEDTAK).firstOrNull() == null) {
-            "Vedtaksbrev finnes allerede på behandling (id=$behandlingId) og kan ikke opprettes på nytt"
-        }
-
-        if (brukerTokenInfo is Saksbehandler) {
-            val kanRedigeres = behandlingService.hentVedtaksbehandlingKanRedigeres(behandlingId, brukerTokenInfo)
-            if (!kanRedigeres) {
-                throw KanIkkeOppretteVedtaksbrev(behandlingId)
-            }
-        }
-
-        return opprettBrev(
-            sakId = sakId,
-            behandlingId = behandlingId,
-            bruker = brukerTokenInfo,
-            brevKodeMapping = brevKodeMapping,
-            brevtype = Brevtype.VEDTAK,
-            brevDataMapping = brevDataMapping,
-        ).first
-    }
-
     suspend fun opprettBrev(
         sakId: SakId,
         behandlingId: UUID?,
@@ -157,11 +125,3 @@ class Brevoppretter(
             adresse = Adresse(adresseType = "", landkode = "", land = ""),
         )
 }
-
-class KanIkkeOppretteVedtaksbrev(
-    behandlingId: UUID,
-) : UgyldigForespoerselException(
-        code = "KAN_IKKE_ENDRE_VEDTAKSBREV",
-        detail = "Statusen til behandlingen tillater ikke at det opprettes vedtaksbrev",
-        meta = mapOf("behandlingId" to behandlingId),
-    )
