+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/vedtaksbrev/VedtaksbrevRoute.kt
@@ -1,4 +1,4 @@
-package no.nav.etterlatte.brev
+package no.nav.etterlatte.brev.vedtaksbrev
 
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.call
@@ -10,6 +10,7 @@ import io.ktor.server.routing.get
 import io.ktor.server.routing.post
 import io.ktor.server.routing.put
 import io.ktor.server.routing.route
+import no.nav.etterlatte.brev.Brevtype
 import no.nav.etterlatte.libs.common.sak.SakId
 import no.nav.etterlatte.libs.ktor.route.BEHANDLINGID_CALL_PARAMETER
 import no.nav.etterlatte.libs.ktor.route.Tilgangssjekker
