+++ apps/etterlatte-brev-api/src/test/kotlin/no/nav/etterlatte/brev/varselbrev/VarselbrevTest.kt
@@ -118,7 +118,6 @@ class VarselbrevTest(
             Brevoppretter(
                 adresseService,
                 brevRepository,
-                behandlingService,
                 innholdTilRedigerbartBrevHenter,
             )
         val pdfGenerator = mockk<PDFGenerator>()
